Inserted the classification algorithm
The function save model and load model is the same of clustering.
The function for the training now use only the DecisionTree classifier and it save the model
Moreover the infer function permit to use the saved model for inference in the infer api.

TODO Insert the test and save the test result in a file
TODO insert the save_path in the config for the test result.