Configuration changes
With these changes the configuration file was updated to last version that permit to have the subdivision of the mapping saved file, the same for the report.
Added the division of the training for each model one the occupation analysis and one for the skill analysis.
Moreover the uri definition for the api was updated with the two models definition for the training and for the inference.